+++ app/src/main/java/com/jonnycaley/cryptomanager/data/model/DataBase/Transaction.kt
@@ -1,9 +1,10 @@
 package com.jonnycaley.cryptomanager.data.model.DataBase
 
 import java.io.Serializable
+import java.math.BigDecimal
 import java.util.*
 
-data class Transaction(var exchange : String, var symbol : String, var pairSymbol : String?, var quantity : Float, var price : Float, var priceUSD : Double, var date : Date, var notes : String?, var isDeducted : Boolean, var isDeductedPriceUsd : Double, var baseImageUrl : String?, var pairImageUrl : String?, var btcPrice: Double, var ethPrice : Double) : Serializable
+data class Transaction(var exchange : String, var symbol : String, var pairSymbol : String?, var quantity : BigDecimal, var price : BigDecimal, var priceUSD : BigDecimal, var date : Date, var notes : String?, var isDeducted : Boolean, var isDeductedPriceUsd : BigDecimal, var baseImageUrl : String?, var pairImageUrl : String?, var btcPrice: BigDecimal, var ethPrice : BigDecimal) : Serializable
 
 //pairSymbol is null for fiat deposits/withdrawls (>0/<0)
 //pairSymbol is not null for crypto holdings e.g. BTC/ETH symbol would be BTC pairSymbol would be ETH ------ buy/sell -> >0/<0
