+++ app/src/main/java/com/jonnycaley/cryptomanager/ui/fiat/FiatContract.kt
@@ -3,15 +3,16 @@ package com.jonnycaley.cryptomanager.ui.fiat
 import com.jonnycaley.cryptomanager.data.model.DataBase.Transaction
 import com.jonnycaley.cryptomanager.utils.mvp.BasePresenter
 import com.jonnycaley.cryptomanager.utils.mvp.BaseView
+import java.math.BigDecimal
 
 interface FiatContract {
     interface View : BaseView<Presenter> {
         fun getFiatCode(): String
         fun showProgressBar()
         fun showTransactions(fiatSymbol : String, showTransactions: List<Transaction>)
-        fun showAvailableFiat(fiatSymbol: String, availableFiatCount: Float)
-        fun showDepositedFiat(fiatSymbol: String, depositedFiatCount: Float)
-        fun showWithdrawnFiat(fiatSymbol: String, withdrawnFiatCount: Float)
+        fun showAvailableFiat(fiatSymbol: String, availableFiatCount: BigDecimal)
+        fun showDepositedFiat(fiatSymbol: String, depositedFiatCount: BigDecimal)
+        fun showWithdrawnFiat(fiatSymbol: String, withdrawnFiatCount: BigDecimal)
     }
 
     interface Presenter : BasePresenter {
